fix: updat wrappers
