+++ apps/web/screens/CustomPage/CustomPageWrapper.tsx
@@ -7,9 +7,13 @@ import {
   CustomPageUniqueIdentifier,
   Query,
   QueryGetCustomPageArgs,
+  QueryGetCustomSubpageArgs,
 } from '@island.is/web/graphql/schema'
 import { useI18n } from '@island.is/web/i18n'
-import { GET_CUSTOM_PAGE_QUERY } from '@island.is/web/screens/queries/CustomPage'
+import {
+  GET_CUSTOM_PAGE_QUERY,
+  GET_CUSTOM_SUBPAGE_QUERY,
+} from '@island.is/web/screens/queries/CustomPage'
 import { Screen, ScreenContext } from '@island.is/web/types'
 
 interface CustomScreenContext extends ScreenContext {
@@ -22,7 +26,7 @@ export type CustomScreen<Props> = ComponentType<
   getProps?: (ctx: CustomScreenContext) => Promise<Props>
 }
 
-export interface CustomPageWrapperProps {
+interface CustomPageWrapperProps {
   customPageData?: CustomPage | null
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   pageProps: any
@@ -89,4 +93,69 @@ export const withCustomPageWrapper = <Props,>(
   }
 
   return CustomPageWrapper
+}
+
+export const withCustomSubpageWrapper = <Props,>(
+  pageSlug: string,
+  parentPageId: CustomPageUniqueIdentifier,
+  Component: CustomScreen<
+    Props & {
+      customSubpageData?: CustomPageWrapperProps['customPageData']
+    }
+  >,
+) => {
+  const CustomSubpageWrapper: Screen<CustomPageWrapperProps> = ({
+    customPageData,
+    pageProps,
+  }) => {
+    console.log(customPageData)
+    console.log(pageProps)
+    const { activeLocale } = useI18n()
+    return (
+      <IntlProvider
+        locale={activeLocale}
+        messages={customPageData?.translationStrings}
+      >
+        <HeadWithSocialSharing
+          title={customPageData?.ogTitle ?? ''}
+          description={customPageData?.ogDescription ?? undefined}
+          imageContentType={customPageData?.ogImage?.contentType}
+          imageWidth={customPageData?.ogImage?.width?.toString()}
+          imageHeight={customPageData?.ogImage?.height?.toString()}
+          imageUrl={customPageData?.ogImage?.url}
+        />
+        <Component {...pageProps} customPageData={customPageData} />
+      </IntlProvider>
+    )
+  }
+
+  CustomSubpageWrapper.getProps = async (context) => {
+    const {
+      data: { getCustomSubpage: customSubpageData },
+    } = await context.apolloClient.query<Query, QueryGetCustomSubpageArgs>({
+      query: GET_CUSTOM_SUBPAGE_QUERY,
+      variables: {
+        input: {
+          lang: context.locale,
+          slug: pageSlug,
+          parentPageId,
+        },
+      },
+    })
+
+    console.log(customSubpageData)
+
+    const [pageProps] = await Promise.all([
+      Component?.getProps?.({ ...context, customPageData: customSubpageData }),
+    ])
+
+    return {
+      customSubpageData,
+      pageProps,
+      customAlertBanner: customSubpageData?.alertBanner,
+      ...pageProps,
+    }
+  }
+
+  return CustomSubpageWrapper
 }
