--- apps/web/screens/CustomPage/index.ts
@@ -1,2 +1,0 @@
-export * from './CustomPageWrapper'
-export * from './CustomSubpageWrapper'
